채팅메시지 문제 해결 안됨

+++ src/main/java/kr/swj/baseball/chat/model/dto/ChatRoomPreview.java
@@ -8,7 +8,7 @@ import lombok.ToString;
 @Setter
 @ToString
 public class ChatRoomPreview {
-    private Long roomNo;
+    private int roomNo;
     private String roomName;
     private String teamLogoUrl;
     private String lastMessage;
